filters: move annotate into own package
This avoids import cycle when GetAnnotations helper is imported from
subpackage of filters.

Followup on #3022
